requirements.txt
pandas
pillow
numpy
matplotlib
openai
seaborn
tensorflow
opencv-python
opencv-python-headless
ultralytics
uvicorn[standard]
gunicorn
fastapi[standard]
loguru
roboflow
streamlit
streamlit-lottie